// backup/features/products/types/types.ts
export type ProductColor = {
  id: string;
  name: string;
  hex: string;
};
// ...
export type ColorsSectionData = {
  colors: ProductColor[];
};
// ...
export type ColorsSectionProps = {
  value?: ColorsSectionData;
  onChange?: (data: ColorsSectionData) => void;
};